chore(): refactor validation and tests
